requirements.txt
text
fastapi==0.95.2
uvicorn==0.22.0
gradio==3.44.0
python-docx==0.8.11
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tqdm==4.66.1
pydantic==1.10.7
requests==2.31.0
python-multipart==0.0.6
pytest==7.3.1